backend/requirements.txt
# AI Engine
openai>=1.0.0
google-generativeai>=0.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
websockets>=12.0
python-multipart>=0.0.9
# Expression detection — Python 3.14 compatible (no TensorFlow needed)
opencv-python>=4.8.0
onnxruntime>=1.18.0
# Audio processing & Wake-word deps
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0
